pyproject.toml
[project]
name = "whisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "openai==0.28",
    "python-dotenv>=1.2.1",
]
